исправил getSalaryAverage()

+++ src/main/java/com/skypro/employee/controller/EmployeeController.java
@@ -36,7 +36,7 @@ public class EmployeeController {
     }
 
     @GetMapping("/employee/salary/average")
-    public int getSalaryAverage() {
+    public double getSalaryAverage() {
         return this.employeeService.getSalaryAverage();
     }
